requirements.txt
openai
sentence-transformers
faiss-cpu
pypdf
pandas
numpy
scikit-learn
python-dotenv
click
pyarrow
fastparquet
flask
gunicorn